requirements.txt
# Audio and Speech
edge-tts>=6.1.3
groq>=0.4.1
SpeechRecognition>=3.10.0
pyaudio>=0.2.11

# Computer Vision and Image Processing
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0

# Deep Learning (Wav2Lip AI)
torch>=1.9.0
torchvision>=0.10.0
librosa>=0.9.0
scikit-image>=0.19.0
scipy>=1.9.0

# Facial Recognition and Landmarks
dlib>=19.22.0
imutils>=0.5.4

# Game/Display Framework
pygame>=2.5.0

# Utility
python-dotenv>=1.0.0

# Note: Install system packages on Raspberry Pi:
# sudo apt update
# sudo apt install cmake build-essential python3-dev
# sudo apt install libasound2-dev portaudio19-dev
# sudo apt install libhdf5-dev libhdf5-serial-dev
# sudo apt install libatlas-base-dev 